Fixed bug in the filtering of the block list and added logging.
The filtering of the private ip ranges was broken in that it did only check
for overlap in one direction which could leave overlaps undetected if the
range coming from the block list is bigger and not starting at the same ip.
A check in both directions should do the trick.

Furthermore, added logging so that it is visible when a block list entry is
ignored due to overlap with a reserved range.